requirements.txt
Flask==0.10.1
requests==2.9.1
Flask-WTF==0.12
python-binary-memcached
mysql-connector-python